backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
pglast==6.3
gunicorn==23.0.0
